+++ maptasker/src/guiwins.py
@@ -437,11 +437,15 @@ class CTkTextview(ctk.CTkFrame):
         self._setup_textbox(master)
 
         # Setup translation text
-        self.translation = {}
-        self.translation["project"] = translate_string("Project: ")
-        self.translation["profile"] = translate_string("Profile: ")
-        self.translation["task"] = translate_string("Task: ")
-        self.translation["scene"] = translate_string("Scene: ")
+        self.translation = {
+            "project": translate_string("Project: "),
+            "profile": translate_string("Profile: "),
+            "task": translate_string("Task: "),
+            "scene": translate_string("Scene: "),
+            "global_variables": translate_string("Project Global Variables"),
+            "directory": translate_string("Directory    (blue entries are hotlinks)"),
+            "properties": translate_string("Properties..."),
+        }
 
         # Process the data and insert it into the text box.
         self.process_data(the_data)
@@ -1818,17 +1822,17 @@ class CTkTextview(ctk.CTkFrame):
               format "Owning <owner_type>: <owner_name>  >>> ".
             - If no match is found, the function returns an empty string.
         """
-        # FIX Translation support
+        # Translation support
         # Global variables are a special case.
         pgv = translate_string("Project Global Variables")
         owner = translate_string("Owner")
         if pgv == line:
             return f"<<< {owner}={pgv}", "", ""
         owner_keys = [
-            "Task: ",
-            "Profile: ",
-            "Project: ",
-            "Scene: ",
+            self.translation("task"),
+            self.translation("profile"),
+            self.translation("project"),
+            self.translation("scene"),
             pgv,
         ]
         invalid_keys = [
@@ -2544,7 +2548,8 @@ class CTkTextview(ctk.CTkFrame):
                 tags,
             )
             # Are we about to do the directory?
-            if text_content[0] == "Directory\n":  # Using text_content here
+            temp = text_content[0]
+            if isinstance(temp, list) and temp[0].startswith("Directory    (blue entries are hotlinks)"):
                 # Handle the hotlink for going up one or more levels.
                 (
                     new_line_num,
@@ -2818,7 +2823,7 @@ class CTkTextview(ctk.CTkFrame):
         for text_linenum, text in enumerate(value["text"]):
             if text == "Directory\n":
                 # Replace text with the formatted directory description
-                value["text"][text_linenum] = ["Directory    (blue entries are hotlinks)\n \n"]
+                value["text"][text_linenum] = [f"{translate_string('Directory    (blue entries are hotlinks)')}\n\n"]
 
             # Process directory headings (e.g. 'Projects...')
             elif text.startswith("\nn"):
@@ -2838,13 +2843,15 @@ class CTkTextview(ctk.CTkFrame):
                 )
                 line_num += 1
 
-                # Restore original text and color
-                value["text"][text_linenum] = (
-                    save_text.replace("Projects...", f"{translate_string('Projects')}...")
-                    .replace("Profiles...", f"{translate_string('Profiles')}...")
-                    .replace("Tasks...", f"{translate_string('Tasks')}...")
-                    .replace("Scenes...", f"{translate_string('Scenes')}...")
-                )
+                # Translate directory headers and restore original text and color
+                if "Projects..." in save_text:
+                    value["text"][text_linenum] = save_text.replace("Projects...", f"{translate_string('Projects')}...")
+                elif "Profiles..." in save_text:
+                    value["text"][text_linenum] = save_text.replace("Profiles...", f"{translate_string('Profiles')}...")
+                elif "Tasks..." in save_text:
+                    value["text"][text_linenum] = save_text.replace("Tasks...", f"{translate_string('Tasks')}...")
+                elif "Scenes..." in save_text:
+                    value["text"][text_linenum] = save_text.replace("Scenes...", f"{translate_string('Scenes')}...")
                 value["color"][text_linenum] = save_color
 
             # Output the updated text and color
@@ -2898,7 +2905,7 @@ class CTkTextview(ctk.CTkFrame):
         hotlink_name = value["directory"][1]
 
         # Determine the name to go up to, which will be used for the tag id.
-        name_to_go_up = hotlink_name if hotlink_name else "entire configuration"
+        name_to_go_up = hotlink_name if hotlink_name else translate_string("entire configuration")
 
         # Check for special "Up One Level" hotlink and modify the text to be displayed if it is.
         up_one_level = False
@@ -2907,7 +2914,7 @@ class CTkTextview(ctk.CTkFrame):
             up_one_level = True
             directory_type = f"{directory_type[2:]}_up"
             object_name = directory_type[:-3].capitalize() if hotlink_name else ""
-            hotlink_name = f"Up One Level to {object_name}: {name_to_go_up}"
+            hotlink_name = f"{translate_string('Up One Level to')} {object_name}: {name_to_go_up}"
             name_to_insert, spacer = hotlink_name, ""
         else:
             # Normal directory entry.  If name greater than spacing allows, truncate it.
@@ -3046,11 +3053,19 @@ class CTkTextview(ctk.CTkFrame):
             self.out_trace_file.write(formatted_message)
 
         # Pre-check for object labels to reduce substring scans.  Force a newline if it is.
-        if any(x in formatted_message for x in ("Task:", "Profile:", "Project:")):
+        if any(
+            x in formatted_message
+            for x in (self.translation["task"], self.translation["profile"], self.translation["project"])
+        ):
             textview_textbox.insert("end", "\n", tag_id)
             line_num += 1
             char_position = 0
             line_num_str = str(line_num)
+
+        # Remove highlighting indicators
+        formatted_message = (
+            formatted_message.replace("<mark>", "").replace("</mark>", "").replace("<b>", "").replace("</b>", "")
+        )
 
         # Insert formatted message and update char position
         char_position = self._insert_message(
@@ -3085,27 +3100,28 @@ class CTkTextview(ctk.CTkFrame):
         message = message.replace("\n\n", "\n").replace("Go to top", "")
 
         # Translate Tasker object names
-        if PrimeItems.program_arguments["language"] not in ("English"):
-            if "Task: " in message:
-                message = message.replace("Task: ", self.translation["task"])
-            elif "Profile: " in message:
-                message = message.replace("Profile: ", self.translation["profile"])
-            elif "Project: " in message:
-                message = message.replace("Project: ", self.translation["project"])
-            elif "Scene: " in message:
-                message = message.replace("Scene: ", self.translation["scene"])
-            elif "Directory\n" in message:
-                message = message.replace("Directory\n", translate_string("Directory") + "\n")
-            elif "Project Global Variables" in message:
-                message = message.replace(
-                    "Project Global Variables",
-                    translate_string("Project Global Variables"),
-                )
-            elif "Properties..." in message:
-                message = message.replace("Properties...", translate_string("Properties..."))
+        # 1. Define the mapping of "Search String" : "Replacement String"
+        # Note: We handle the mixed sources (self.translation vs translate_string) here.
+        replacement_map = {
+            "Task: ": self.translation["task"],
+            "Profile: ": self.translation["profile"],
+            "Project: ": self.translation["project"],
+            "Scene: ": self.translation["scene"],
+            "Directory    ": self.translation["directory"] + "\n",
+            "Project Global Variables": self.translation["global_variables"],
+            "Properties...": self.translation["properties"],
+        }
+
+        # 2. Iterate through the map and apply replacement
+        for search_term, replacement in replacement_map.items():
+            if search_term in message:
+                message = message.replace(search_term, replacement)
+                # Uncomment the line below if you strictly want the original 'elif' behavior
+        # (stop looking after the first match is found):
+        # break
 
         # Handle special case for 'directory'
-        if previous_value == "directory" and translate_string("Project: ") in message:
+        if previous_value == "directory" and self.translation["project"] in message:
             message = f"\n{message}"
 
         # # Short-circuit for empty messages
@@ -3370,7 +3386,7 @@ class CTkTextview(ctk.CTkFrame):
         # starting_line_to_search = 1
         # Don't addf highlight if this is a label.  We've already added it.
         with contextlib.suppress(KeyError):
-            if text_linenum == 0 and value["highlights"] and "-text" not in tag_id:
+            if text_linenum == 0 and value["highlights"] and value["highlights"][0] and "-text" not in tag_id:
                 tags = self.add_highlights(message, value, text_linenum, previous_value, tag_id, tags)
 
         # Now color the text.
@@ -3420,12 +3436,12 @@ class CTkTextview(ctk.CTkFrame):
             "h5-text": {"font": "h5"},
             "h6-text": {"font": "h6"},
         }
-
+        translation = self.textview_textbox.master.translation
         search_word_mapping = {
-            "Task: ": "Task: ",
-            "Profile: ": "Profile: ",
-            "Project: ": "Project: ",
-            "Scene: ": "Scene: ",
+            "Task: ": translation["task"],
+            "Profile: ": translation["profile"],
+            "Project: ": translation["project"],
+            "Scene: ": translation["scene"],
         }
 
         # # Find the search word context.  Default to text in message if not found.
@@ -3433,6 +3449,7 @@ class CTkTextview(ctk.CTkFrame):
             (word for word in search_word_mapping if word in message),
             message,
         )
+
         # Get the highlight
         highlight = value["highlights"][text_linenum]
         if highlight: